+++ include/tilck/kernel/arch/i386/arch_utils.h
@@ -39,7 +39,7 @@ typedef struct {
 
 static ALWAYS_INLINE int regs_intnum(regs *r)
 {
-   return r->int_num;
+   return (int)r->int_num;
 }
 
 static ALWAYS_INLINE void set_return_register(regs *r, u32 value)
